# [3.6] Reduce performance regression in RSA public operations