… ChangeLog.d/fix-rsa-performance-regression.txt
… library/bignum_internal.h
… library/rsa.c
… tests/suites/test_suite_bignum.function
… tests/suites/test_suite_bignum_core.function