## Description

Fix #9232 - partially, the low-hanging fruits (also those that don't make the code size go up by too much).

Status: work in progress.

TODO:
- [x] measure code size
- [x] measure perf on other platforms
- [x] currently the window size is based on the limbsize of E, should it use the actual bitsize instead when E is public?
- [x] decide on API
- [x] propagate to RSA functions

## PR checklist

- [x] **changelog** provided
- [ ] **development** TODO
- [ ] **3.6 backport** this is it
- [x] **2.28 backport** not required (we are fixing a regression in 3.6)
- [ ] **tests** partly provided, will be continued in https://github.com/Mbed-TLS/mbedtls/pull/9493