@mpg 2024-07-22
> Ok (with a few minor remarks) if we really have to have mixed-leakiness functions, but do we really have to do that?

No, I don't think we absolutely have to do that, and that's one of the main points about which I was hoping to get feedback. I just has to make a temporary choice for prototyping.

> One opinion about the API: maybe it would be worth to keep the `_optionally_safe()` function static and have an `_unsafe()` version in the header instead.

Yes, I think that would already be better.

> I don't like the complexity of having functions whose security properties depend on a runtime argument.

Agreed. Do you think it's acceptable for `static` functions though or would you rather avoid it entirely?

> The only argument I can think of to merge the two is code size. Is the gain really worth it?

That, and also there can be a maintenance cost if we end up duplicating source code. I can give it a try to do some size measurements and check how easy or awkward it is to avoid duplication.