@gilles-peskine-arm 2024-07-22
> > I don't like the complexity of having functions whose security properties depend on a runtime argument.
> 
> Agreed. Do you think it's acceptable for static functions though or would you rather avoid it entirely?

I'd prefer to avoid it, but for smaller functions with a limited scope, it's not so bad.