@jforissier 2024-07-23
With this PR (and the necessary changes to make use of the `E_public` parameter), the OP-TEE test case runs faster although still much slower than with v3.5.2. This is because the test doesn't only run RSA signature verification and decryption, it also includes RSA signing and encryption. Assuming OP-TEE switches to the more secure algorithm for secret stuff, we may need to disable this test in the CI and run it only during the release tests.
Here are the numbers (running `time xtest 4011` on the arm32 QEMU target):

- v3.5.2 (optee_os master @ b339ffbd9): 14.5s
- v3.6.0: 8m 26.6s
- This PR (my branch: [test-mbedtls-expmod-regression-fix-9281](https://github.com/jforissier/optee_os/tree/test-mbedtls-expmod-regression-fix-9281) @ 9dca1658f): 5m 58s