@tom-cosgrove-arm 2024-08-14
> If we're going to have a single function for both cases, I would prefer a more specific naming convention than `optionally_safe`. I don't like the use of “safe”, which could refer to many different things. I'd prefer [leaky](https://github.com/Mbed-TLS/mbedtls/blob/mbedtls-3.6.0/3rdparty/p256-m/p256-m/p256-m.c#L976) or “public”.

How about using `ct` instead of `safe`?

We could also put the emphasis in the other direction: `can_skip_ct`.