@yanesca 2024-08-15
> How about using ct instead of safe?
>
> We could also put the emphasis in the other direction: can_skip_ct.

I chose the prefix following the existing `mbedtls_mpi_core_get_mont_r2_unsafe()` funtion. It didn't occur to me to tie it to the _ct prefix. Conceptually this seems neat and `leaky` is more precise, but if I think about it I probably like `_unsafe` better:
- In the context of `bignum_core` we don't always use `_ct` but even where we do will go away over time as being constant time will become default.
- I think it is more likely that people will consult the documentation and start to think if they read `optionally_unsafe` as opposed to `can_skip_ct` or `optionally_leaky`

What do you think?