@tom-cosgrove-arm 2024-08-22
Performance measurements, in default config (desktop only, sorry)

aarch64:
|                          Test | base:mbedtls-3.6:28cdd1190 | pr:9281:rebased:dd9609174 | Relative |
|------------------------------:|---------------------------:|---------------------------:|---------:|
| 1024/1024-bit exp_mod, cold r |               0.449740 ms |               0.445420 ms |    1.01x |
| 1024/1024-bit exp_mod,  hot r |               0.444260 ms |               0.442480 ms |    1.00x |
|   17/1024-bit exp_mod, cold r |               0.044070 ms |               0.014440 ms |    3.05x |
|   17/1024-bit exp_mod,  hot r |               0.041390 ms |               0.011440 ms |    3.62x |

x86_64:
|                          Test | base:mbedtls-3.6:28cdd1190 | pr:9281:rebased:23675570a | Relative |
|------------------------------:|---------------------------:|---------------------------:|---------:|
| 1024/1024-bit exp_mod, cold r |               0.672410 ms |               0.653440 ms |    1.03x |
| 1024/1024-bit exp_mod,  hot r |               0.670990 ms |               0.649560 ms |    1.03x |
|   17/1024-bit exp_mod, cold r |               0.066270 ms |               0.020480 ms |    3.24x |
|   17/1024-bit exp_mod,  hot r |               0.062450 ms |               0.016760 ms |    3.73x |